+++ ontol-compiler/tests/compiler-integration/errors/test_compile_errors.rs
@@ -141,7 +141,7 @@ fn non_disjoint_string_union() {
 fn sequence_mix1() {
     "
     type u {
-        rel .is?: int // ERROR invalid mix of relationship type for subject
+        rel .is?: i64 // ERROR invalid mix of relationship type for subject
         rel .0: string
     }
     "
